+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-2.1.2...master
 
+### Fixed
+- Only one thread can update `Throttled Cache` values
+
 ## [2.1.2] - 2021-07-01
 [2.1.2]: https://github.com/atlassian-labs/db-replica/compare/release-2.1.0...release-2.1.2